avoid error if Spotify not installed
sending a tell to a nonexistent application bombs out in AppleScript,
so need to make sure application exists at all before checking to see
if it’s running.

iTunes should always be there since we’re assuming a macOS system for
now, but might as well test for it too for future-proofing, even if
verbose in current methodology.

♬ : The Clash / Lost In the Supermarket